+++ lib/base-chips/datastore.mongo.js
@@ -22,6 +22,9 @@ Sealious.ConfigManager.set_default_config(
 
 datastore_mongo.start = function(){
 	var config = Sealious.ConfigManager.get_config("datastore_mongo");
+
+	console.log("config.embedded", config.embedded);
+
 	if(config.embedded){
 		TingoDB = require('tingodb')();
