+++ lib/core-services/resource-manager.js
@@ -35,7 +35,6 @@ var UUIDGenerator = require("uid");
 				body: encoded_body,
 				created_context: context.toObject(),
 			};
-			console.log("create resource_data", resource_data);
 			return Sealious.Dispatcher.datastore.insert("resources", resource_data, {});
 		}).then(function(inserted_document){
 			var database_entry = inserted_document;
@@ -247,6 +246,118 @@ var UUIDGenerator = require("uid");
  				reject(new Sealious.Errors.Error("ResourceManager tried to access resource type `" + type_name + "`, which does not exist."));
  			}
  		})
+ 	}
+
+ 	this.test_init = function(){
+ 		// Creating resource always fails
+ 		var always_fails = new Sealious.ChipTypes.FieldType("always_fails");	
+ 		always_fails.prototype.isProperValue = function(value_in_code){
+ 			return new Promise(function(resolve, reject){
+ 				reject();
+ 			})
+ 		}
+ 		var always_fails_resource = new Sealious.ChipTypes.ResourceType({
+ 			name: "always_fails_resource",
+ 			fields: [
+ 				{name: "#fail", type: "always_fails"}
+ 			]
+ 		});
+
+
+ 		//Creating resource never fails
+ 		var never_fails = new Sealious.ChipTypes.FieldType("never_fails");
+ 		never_fails.prototype.isProperValue = function(value_in_code){
+ 			return new Promise(function(resolve,reject){
+ 				resolve();
+ 			})
+ 		}
+
+ 		var never_fails_resource = new Sealious.ChipTypes.ResourceType({
+ 			name: "never_fails_resource",
+ 			fields: [
+ 				{name: "#success", type: "never_fails"},
+ 			]
+ 		});
+ 	}
+
+ 	this.test_start = function(){
+ 		describe("ResourceManager", function(){
+			var context = Sealious.Context();
+
+		    it('should not create a new resource', function(done){
+				Sealious.Dispatcher.resources.create(context, "always_fails_resource", { "#fail": "tak" })
+				.then(function(){
+					done(new Error("It didn't throw an error!"));
+				}).catch(function(error){
+					done();
+				});
+		 	});
+		 	it('should create a new resource', function(done){
+				Sealious.Dispatcher.resources.create(context, "never_fails_resource", { "#success": "tak" })
+				.then(function(result){
+					done();
+				}).catch(function(error){
+					done(error);
+				});
+		 	});
+		 	it('should get data about the resource', function(done){
+				Sealious.Dispatcher.resources.create(context, "never_fails_resource", { "#success": "tak" })
+				.then(function(result){
+					return Sealious.Dispatcher.resources.get_by_id(context, result.id)
+					.then(function(resource){
+						done();	
+					})
+				}).catch(function(error){
+					done(error);
+				});
+		 	});
+		 	it('should delete the resource', function(done){
+				Sealious.Dispatcher.resources.create(context, "never_fails_resource", { "#success": "tak" })
+				.then(function(result){
+					Sealious.Dispatcher.resources.delete(context, "never_fails_resource", result.id)
+					.then(function(resource){
+						done();	
+					}).catch(function(error){
+						done(new Error(error));
+					})
+				}).catch(function(error){
+					done(new Error("It threw an error!"));
+				});
+		 	});
+		 	it('should update the resource', function(done){
+				Sealious.Dispatcher.resources.create(context, "never_fails_resource", { "#success": "tak" })
+				.then(function(result){
+					Sealious.Dispatcher.resources.update_resource(context, "never_fails_resource", result.id, {"#success" : "tak2"})
+					.then(function(resource){
+						done();	
+					}).catch(function(error){
+						done(new Error(error));
+					})
+				}).catch(function(error){
+					done(error);
+				});
+		 	});
+		 	it('should list resources by type', function(done){
+				Sealious.Dispatcher.resources.list_by_type(context, "never_fails_resource")
+				.then(function(result){
+					done();
+				}).catch(function(error){
+					done(new Error("It threw an error!"));
+				});
+		 	});
+		 	it('should get resource type signature (schema)', function(done){
+				Sealious.Dispatcher.resources.get_resource_type_signature(context, "never_fails_resource")
+				.then(function(result){
+					done();
+				}).catch(function(error){
+					done(new Error("It threw an error!"));
+				});
+		 	});
+
+
+		 	//run(); //!!important	
+ 		})
+
  	}
  }
