+++ lib/main.js
@@ -51,4 +51,16 @@ Sealious.start = function(){
 	return Sealious.ChipManager.start_chips(chip_types_to_start);
 }
 
+Sealious.test = function(){
+	var ResourceManager = Sealious.ResourceManager;
+
+	Sealious.init();
+	ResourceManager.test_init();
+
+	Sealious.start().then(function(){
+		ResourceManager.test_start();
+		run(); //function created by Mocha. Present only if Mocha is run with --delay option from the command line
+	})
+}
+
 module.exports = Sealious;
