+++ package.json
@@ -5,7 +5,8 @@
   "description": "",
   "main": "./lib/main.js",
   "scripts": {
-    "test": "multi='dot=- html-cov=coverage.html' ./node_modules/mocha/bin/mocha -r blanket --reporter mocha-multi --delay ./tests/test.js",
+    "test": "./node_modules/mocha/bin/mocha --delay ./tests/test.js",
+    "test-coverage": "./node_modules/mocha/bin/mocha -R html-cov --delay ./tests/test.js >> coverage.html && xdg-open coverage.html",
     "test-no-blanket": "mocha --delay ./tests/test.js"
   },
   "repository": {
